Progress.

+++ include/mpi/core/environment.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
 #include <string>
 
@@ -63,7 +64,7 @@ inline std::string    processor_name      ()
   std::string  result(MPI_MAX_PROCESSOR_NAME, ' ');
   std::int32_t size  (0);
   MPI_CHECK_ERROR_CODE(MPI_Get_processor_name, (&result[0], &size))
-  result.resize(size);
+  result.resize(static_cast<std::size_t>(size));
   return result;
 }
